requirements.txt
asgiref==3.7.2
attrs==23.1.0
django-filter==23.3
django-resized==1.0.2
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema-specifications==2023.7.1
jsonschema==4.19.1
Markdown==3.4.4
Pillow==10.0.1
pip==22.3.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
setuptools==65.5.1
sorl-thumbnail==12.9.0
sqlparse==0.4.4
uritemplate==4.1.1
wheel==0.38.4
django-environ==0.11.2